Added home page and minor code refractoring
The home page was added to the application including the filtering option. Router added in Controller Component and the feature when user click on any released movie it directs to the details page. Minor code refractoring in the Authenticator component and the react-router-dom version upgraded to latest because the useParams and useHistory were not getting imported